requirements.txt
Flask
vercel-python
ultralytics
opencv-python
numpy 
python-dotenv 